Modularize the inquisition importer.
Makeit possible to import questions without importing a quiz.
